% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desencriptar_aes.R
\name{desencriptar_aes}
\alias{desencriptar_aes}
\title{Desencripta en AES CBC}
\usage{
desencriptar_aes(dato, modo = "CBC", vi, clave)
}
\arguments{
\item{dato, }{modo, vi, clave}
}
\value{
json
}
\description{
Desencripta en AES CBC un dato encriptado y en formato base64. La clave y el vi deben de ir en formato string sin espacios
}
\examples{
 desencriptar_aes("ANOZ8RrW8zGvcqKEXJEe8A==", "CBC", "47,148,14,86,131,53,108,190,252,83,71,46,69,201,12,119" , "01,02,03,04,05,06,07,08,09,0a,0b,0c,0d,0e,0f,10")

}
